+++ KSP2Runtime/KSPVessel/KSPVesselModule.ActionGroups.cs
@@ -59,7 +59,7 @@ namespace KontrolSystem.KSP.Runtime.KSPVessel {
                 get => vessel.GetActionGroupState(KSPActionGroup.Abort) == KSPActionGroupState.True;
                 set => vessel.SetActionGroup(KSPActionGroup.Abort, value);
             }
-            
+
             [KSField]
             public bool Custom1 {
                 get => vessel.GetActionGroupState(KSPActionGroup.Custom01) == KSPActionGroupState.True;
@@ -77,43 +77,43 @@ namespace KontrolSystem.KSP.Runtime.KSPVessel {
                 get => vessel.GetActionGroupState(KSPActionGroup.Custom03) == KSPActionGroupState.True;
                 set => vessel.SetActionGroup(KSPActionGroup.Custom03, value);
             }
-            
+
             [KSField]
             public bool Custom4 {
                 get => vessel.GetActionGroupState(KSPActionGroup.Custom04) == KSPActionGroupState.True;
                 set => vessel.SetActionGroup(KSPActionGroup.Custom04, value);
             }
-            
+
             [KSField]
             public bool Custom5 {
                 get => vessel.GetActionGroupState(KSPActionGroup.Custom05) == KSPActionGroupState.True;
                 set => vessel.SetActionGroup(KSPActionGroup.Custom05, value);
             }
-            
+
             [KSField]
             public bool Custom6 {
                 get => vessel.GetActionGroupState(KSPActionGroup.Custom06) == KSPActionGroupState.True;
                 set => vessel.SetActionGroup(KSPActionGroup.Custom06, value);
             }
-            
+
             [KSField]
             public bool Custom7 {
                 get => vessel.GetActionGroupState(KSPActionGroup.Custom07) == KSPActionGroupState.True;
                 set => vessel.SetActionGroup(KSPActionGroup.Custom07, value);
             }
-            
+
             [KSField]
             public bool Custom8 {
                 get => vessel.GetActionGroupState(KSPActionGroup.Custom08) == KSPActionGroupState.True;
                 set => vessel.SetActionGroup(KSPActionGroup.Custom08, value);
             }
-            
+
             [KSField]
             public bool Custom9 {
                 get => vessel.GetActionGroupState(KSPActionGroup.Custom09) == KSPActionGroupState.True;
                 set => vessel.SetActionGroup(KSPActionGroup.Custom09, value);
             }
-            
+
             [KSField]
             public bool Custom10 {
                 get => vessel.GetActionGroupState(KSPActionGroup.Custom10) == KSPActionGroupState.True;