+++ TO2/AST/Call.cs
@@ -219,10 +219,20 @@ namespace KontrolSystem.TO2.AST {
                 return;
             }
 
+            if (arguments.Count > function.Parameters.Count) {
+                context.AddError(new StructuralError(
+                    StructuralError.ErrorType.ArgumentMismatch,
+                    $"Function '{FullName}' only allows {function.Parameters.Count} arguments, {arguments.Count} where given",
+                    Start,
+                    End
+                ));
+                return;
+            }
+
             if (function.RequiredParameterCount() > arguments.Count) {
                 context.AddError(new StructuralError(
                     StructuralError.ErrorType.ArgumentMismatch,
-                    $"Function '{FullName}' requires at least {function.RequiredParameterCount()} arguments",
+                    $"Function '{FullName}' requires at least {function.RequiredParameterCount()} arguments, {arguments.Count} where given",
                     Start,
                     End
                 ));